+++ core/navigation/src/main/java/com/teamoffroad/core/navigation/RouteModel.kt
@@ -24,9 +24,6 @@ sealed interface MainTabRoute : Route {
 
 sealed interface AuthRoute : Route {
     @Serializable
-    data object KaKaoLogin: AuthRoute
-
-    @Serializable
     data object SetNickname : AuthRoute
 
     @Serializable
