+++ feature/auth/src/main/java/com/teamoffroad/feature/auth/navigation/AuthNavigation.kt
@@ -10,7 +10,6 @@ import com.teamoffroad.core.navigation.AuthRoute
 import com.teamoffroad.core.navigation.Route
 import com.teamoffroad.feature.auth.presentation.AgreeTermsAndConditionsScreen
 import com.teamoffroad.feature.auth.presentation.AuthScreen
-import com.teamoffroad.feature.auth.presentation.KakaoLoginScreen
 import com.teamoffroad.feature.auth.presentation.SelectedCharacterScreen
 import com.teamoffroad.feature.auth.presentation.SetBirthDateScreen
 import com.teamoffroad.feature.auth.presentation.SetCharacterScreen
@@ -19,10 +18,6 @@ import com.teamoffroad.feature.auth.presentation.SetNicknameScreen
 import java.net.URLDecoder
 import java.net.URLEncoder
 import java.nio.charset.StandardCharsets
-
-fun NavController.navigateToKakaoLogin() {
-    navigate(AuthRoute.KaKaoLogin)
-}
 
 fun NavController.navigateToAgreeTermsAndConditions() {
     navigate(AuthRoute.AgreeTermsAndConditions)
@@ -56,7 +51,6 @@ fun NavController.navigateToSelectedCharacter(
 
 @RequiresApi(Build.VERSION_CODES.TIRAMISU)
 fun NavGraphBuilder.authNavGraph(
-    navigateToKakaoLogin : () -> Unit,
     navigateToHome: () -> Unit,
     navigateToAgreeTermsAndConditions: () -> Unit,
     navigateToSetNickname: () -> Unit,
@@ -70,12 +64,6 @@ fun NavGraphBuilder.authNavGraph(
         AuthScreen(
             navigateToHome,
             navigateToAgreeTermsAndConditions,
-            navigateToKakaoLogin,
-        )
-    }
-    composable<AuthRoute.KaKaoLogin> {
-        KakaoLoginScreen(
-            navigateToBack
         )
     }
     composable<AuthRoute.AgreeTermsAndConditions> {
