+++ feature/auth/src/main/java/com/teamoffroad/feature/auth/presentation/AuthScreen.kt
@@ -1,6 +1,10 @@
 package com.teamoffroad.feature.auth.presentation
 
 import android.app.Activity
+import android.util.Log
+import android.webkit.WebResourceRequest
+import android.webkit.WebView
+import android.webkit.WebViewClient
 import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.foundation.Image
@@ -23,6 +27,7 @@ import androidx.compose.ui.graphics.painter.Painter
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.viewinterop.AndroidView
 import androidx.constraintlayout.compose.ConstraintLayout
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
@@ -34,6 +39,7 @@ import com.teamoffroad.core.designsystem.theme.Main1
 import com.teamoffroad.core.designsystem.theme.Main2
 import com.teamoffroad.core.designsystem.theme.OffroadTheme
 import com.teamoffroad.core.designsystem.theme.White
+import com.teamoffroad.offroad.feature.auth.BuildConfig
 import com.teamoffroad.offroad.feature.auth.R
 import kotlinx.coroutines.delay
 
@@ -41,7 +47,6 @@ import kotlinx.coroutines.delay
 internal fun AuthScreen(
     navigateToHome: () -> Unit,
     navigateToAgreeTermsAndConditions: () -> Unit,
-    navigateToKakaoLogin: () -> Unit,
     viewModel: AuthViewModel = hiltViewModel(),
 ) {
     val isSignInSuccess by viewModel.successSignIn.collectAsStateWithLifecycle()
@@ -56,6 +61,8 @@ internal fun AuthScreen(
             viewModel.performGoogleSignIn(task)
         }
     }
+    var showWebView by remember { mutableStateOf(false) }
+
     LaunchedEffect(Unit) {
         signInLauncherInitialized = true
     }
@@ -98,7 +105,7 @@ internal fun AuthScreen(
                 painter = painterResource(id = R.drawable.ic_auth_kakao_logo),
                 background = Kakao,
                 contentDescription = "auth_kakao",
-                onClick = { navigateToKakaoLogin() },
+                onClick = { showWebView = true },
                 modifier = Modifier.constrainAs(kakaoLogin) {
                     start.linkTo(parent.start, margin = 24.dp)
                     end.linkTo(parent.end, margin = 24.dp)
@@ -123,6 +130,19 @@ internal fun AuthScreen(
         }
     }
     SplashScreen()
+    if (showWebView) {
+        StartKakaoLoginWebView(
+            clientId = BuildConfig.KAKO_CLIENT_ID,
+            redirectUri = BuildConfig.KAKO_REDIRECT_URI,
+            onCodeReceived = { code ->
+                viewModel.performKakaoSignIn(code)
+                Log.d("asdasd", code)
+            },
+            onClose = {
+                showWebView = false
+            },
+        )
+    }
 }
 
 @Composable
@@ -171,4 +191,62 @@ fun ClickableImage(
             )
         }
     }
+}
+
+@Composable
+private fun StartKakaoLoginWebView(
+    clientId: String,
+    redirectUri: String,
+    onCodeReceived: (String) -> Unit,
+    onClose: () -> Unit,
+) {
+    val loginUrl =
+        "https://kauth.kakao.com/oauth/authorize?response_type=code&client_id=$clientId&redirect_uri=$redirectUri"
+
+    AndroidView(
+        modifier = Modifier.fillMaxSize(),
+        factory = { context ->
+            WebView(context).apply {
+                settings.javaScriptEnabled = true
+                webViewClient = object : WebViewClient() {
+                    override fun onPageStarted(
+                        view: WebView?,
+                        url: String?,
+                        favicon: android.graphics.Bitmap?
+                    ) {
+                        super.onPageStarted(view, url, favicon)
+
+                        url?.let {
+                            if (it.startsWith(redirectUri)) {
+                                val uri = android.net.Uri.parse(it)
+                                val code = uri.getQueryParameter("code")
+                                if (code != null) {
+                                    onCodeReceived(code)
+                                    onClose()
+                                }
+                            }
+                        }
+                    }
+
+                    override fun shouldOverrideUrlLoading(
+                        view: WebView?,
+                        request: WebResourceRequest?
+                    ): Boolean {
+                        val url = request?.url.toString()
+                        if (url.startsWith(redirectUri)) {
+                            val uri = android.net.Uri.parse(url)
+                            val code = uri.getQueryParameter("code")
+                            if (code != null) {
+                                onCodeReceived(code)
+                                onClose()
+                                return true
+                            }
+                        }
+                        return super.shouldOverrideUrlLoading(view, request)
+                    }
+                }
+                loadUrl(loginUrl)
+            }
+        }
+    )
 }
