+++ feature/main/src/main/java/com/teamoffroad/feature/main/component/MainNavHost.kt
@@ -105,7 +105,6 @@ internal fun MainNavHost(
                     navigator.navigateToSelectedCharacter(selectedCharacterUrl)
                 },
                 navigateToBack = navigator::popBackStackIfNotMainTabRoute,
-                navigateToKakaoLogin = navigator::navigateToKakaoLogin
             )
         }
     }
